requirements.txt
certifi==2022.12.7
charset-normalizer==3.0.1
idna==3.4
python-dotenv==0.21.1
requests==2.28.2
urllib3==1.26.14
aiohttp==3.8.5
